feat: ライクリポジトリにユーザー取得のメソッドを定義していたのを削除
