+++ reimi/backend/reimi_app/src/main/java/com/reimi/reimi_app/domain/repository/LikeRepository.java
@@ -1,15 +1,11 @@
 package com.reimi.reimi_app.domain.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import com.reimi.reimi_app.domain.model.like.Like;
-import com.reimi.reimi_app.domain.model.user.User;
 import com.reimi.reimi_app.domain.model.user.UserId;
 
 public interface LikeRepository {
-
-    Optional<User> findUserByUserId(UserId userId);
 
     boolean exists(UserId fromUserId, UserId toUserId);
